Lab 7 Authorization

+++ mean-course/backend/routes/post.js
@@ -36,7 +36,8 @@ router.post("", checkAuth, multer({ storage: storage }).single("image"),
         const post = new Post({
             title: req.body.title,
             content: req.body.content,
-            imagePath: url + "/images/" + req.file.filename
+            imagePath: url + "/images/" + req.file.filename,
+            creator: req.userData.userId
         });
         post.save().then(createdPost => {
             res.status(201).json({
@@ -49,6 +50,7 @@ router.post("", checkAuth, multer({ storage: storage }).single("image"),
         });
     }
 );
+
 
 
 router.get("", (req, res, next) => {
@@ -130,26 +132,43 @@ router.put("/:id", checkAuth, multer({ storage: storage }).single("image"),
             _id: req.body.id,
             title: req.body.title,
             content: req.body.content,
-            imagePath: imagePath
+            imagePath: imagePath,
+            creator: req.userData.userId
         });
-        // console.log(post);
-        Post.updateOne({ _id: req.params.id }, post).then(result => {
-            res.status(200).json({ message: "Update successful!" });
+        // Post.updateOne({ _id: req.params.id }, post).then(result => {
+        Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+            console.log(result)
+            if (result.nModified > 0) {
+                res.status(200).json({ message: "Update successful!" });
+            } else {
+                res.status(401).json({ message: "Not authorized!" });
+            }
         });
+
     }
 );
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-    Post.deleteOne({ _id: req.params.id }).then((result) => {
-        console.log(result);
-        res.status(200).json({ message: "Post deleted!" });
-    });
+    Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(
+        result => {
+            console.log(result);
+            if (result.n > 0) {
+                res.status(200).json({ message: "Deletion successful!" });
+            } else {
+                res.status(401).json({ message: "Not authorized!" });
+            }
+        }
+    )
 });
 
 router.delete("", checkAuth, (req, res, next) => {
-    Post.deleteMany().then((result) => {
+    Post.deleteMany({ creator: req.userData.userId }).then((result) => {
         console.log(result);
-        res.status(200).json({ message: "All Post deleted!" });
+        if (result.n > 0) {
+            res.status(200).json({ message: "Delete All successful!" });
+        } else {
+            res.status(401).json({ message: "Not authorized!" });
+        }
     });
 });
